dags/requirements.txt
prometheus-client==0.20.0
requests-aws4auth==1.2.3
pyyaml>=6.0
